+++ lib/tartarus/remote_storage.rb
@@ -1,2 +1,4 @@
 module Tartarus::RemoteStorage
+  autoload :Glacier, "tartarus/remote_storage/glacier"
+  autoload :Null, "tartarus/remote_storage/null"
 end